+++ doc/whats_new.rst
@@ -71,6 +71,8 @@ Enhancements
 - Added some useful properties and methods to the ``xarray.Dataset.xsimlab``
   extension (:issue:`103`).
 - Save model inputs/outputs using zarr (:issue:`102`).
+- Added :class:`~xsimlab.progress.ProgressBar` to track simulation progress
+  (:issue:`104`).
 
 Bug fixes
 ~~~~~~~~~